refactor: Extract manager setup logic into modular providers
Moved dependency initialization and configuration code from server.py
into separate provider modules under `providers/` directory for better
separation of concerns and maintainability.

Each provider module handles a specific aspect of manager setup:
- Database connections and models
- Redis clients (image, live, stat, stream, etc.)
- Event system (dispatcher, producer, hub)
- Background tasks and message queues
- Service discovery and leader election
- Plugin systems (event dispatcher, hooks, network)
- Monitoring and idle checking
- Core repositories and services
- Sokovan orchestrator
