+++ src/ai/backend/manager/providers/__init__.py
@@ -1,0 +1,1 @@
+"""Manager setup context providers."""
