+++ src/ai/backend/manager/server.py
@@ -20,7 +20,6 @@ from collections.abc import (
 )
 from contextlib import asynccontextmanager as actxmgr
 from dataclasses import dataclass
-from datetime import datetime
 from pathlib import Path
 from pprint import pformat
 from typing import (
@@ -41,44 +40,12 @@ from aiohttp import web
 from aiohttp.typedefs import Handler, Middleware
 from setproctitle import setproctitle
 
-from ai.backend.common import redis_helper
-from ai.backend.common.auth import PublicKey, SecretKey
-from ai.backend.common.bgtask.bgtask import BackgroundTaskManager
 from ai.backend.common.cli import LazyGroup
-from ai.backend.common.clients.valkey_client.valkey_bgtask.client import ValkeyBgtaskClient
-from ai.backend.common.clients.valkey_client.valkey_container_log.client import (
-    ValkeyContainerLogClient,
-)
-from ai.backend.common.clients.valkey_client.valkey_image.client import ValkeyImageClient
-from ai.backend.common.clients.valkey_client.valkey_live.client import ValkeyLiveClient
-from ai.backend.common.clients.valkey_client.valkey_schedule.client import ValkeyScheduleClient
-from ai.backend.common.clients.valkey_client.valkey_stat.client import ValkeyStatClient
-from ai.backend.common.clients.valkey_client.valkey_stream.client import ValkeyStreamClient
 from ai.backend.common.config import find_config_file
 from ai.backend.common.data.config.types import EtcdConfigData
-from ai.backend.common.defs import (
-    REDIS_BGTASK_DB,
-    REDIS_CONTAINER_LOG,
-    REDIS_IMAGE_DB,
-    REDIS_LIVE_DB,
-    REDIS_STATISTICS_DB,
-    REDIS_STREAM_DB,
-    REDIS_STREAM_LOCK,
-    RedisRole,
-)
 from ai.backend.common.etcd import AsyncEtcd
-from ai.backend.common.events.dispatcher import EventDispatcher, EventProducer
-from ai.backend.common.events.event_types.artifact_registry.anycast import (
-    DoScanReservoirRegistryEvent,
-)
-from ai.backend.common.events.fetcher import EventFetcher
-from ai.backend.common.events.hub.hub import EventHub
 from ai.backend.common.exception import BackendAIError, ErrorCode
 from ai.backend.common.json import dump_json_str
-from ai.backend.common.leader.tasks.event_task import EventTaskSpec
-from ai.backend.common.message_queue.hiredis_queue import HiRedisQueue
-from ai.backend.common.message_queue.queue import AbstractMessageQueue
-from ai.backend.common.message_queue.redis_queue import RedisMQArgs, RedisQueue
 from ai.backend.common.metrics.http import (
     build_api_metric_middleware,
     build_prometheus_metrics_handler,
@@ -87,30 +54,12 @@ from ai.backend.common.metrics.metric import CommonMetricRegistry
 from ai.backend.common.metrics.profiler import Profiler, PyroscopeArgs
 from ai.backend.common.middlewares.request_id import request_id_middleware
 from ai.backend.common.msgpack import DEFAULT_PACK_OPTS, DEFAULT_UNPACK_OPTS
-from ai.backend.common.plugin.event import EventDispatcherPluginContext
-from ai.backend.common.plugin.hook import ALL_COMPLETED, PASSED, HookPluginContext
 from ai.backend.common.plugin.monitor import INCREMENT
-from ai.backend.common.service_discovery.etcd_discovery.service_discovery import (
-    ETCDServiceDiscovery,
-    ETCDServiceDiscoveryArgs,
-)
-from ai.backend.common.service_discovery.redis_discovery.service_discovery import (
-    RedisServiceDiscovery,
-    RedisServiceDiscoveryArgs,
-)
-from ai.backend.common.service_discovery.service_discovery import (
-    ServiceDiscoveryLoop,
-    ServiceEndpoint,
-    ServiceMetadata,
-)
 from ai.backend.common.types import (
-    AGENTID_MANAGER,
     AgentSelectionStrategy,
-    ServiceDiscoveryType,
 )
 from ai.backend.common.utils import env_info
 from ai.backend.logging import BraceStyleAdapter, Logger, LogLevel
-from ai.backend.logging.otel import OpenTelemetrySpec
 from ai.backend.manager.config.bootstrap import BootstrapConfig
 from ai.backend.manager.config.loader.config_overrider import ConfigOverrider
 from ai.backend.manager.config.loader.etcd_loader import (
@@ -126,22 +75,34 @@ from ai.backend.manager.config.loader.toml_loader import TomlConfigLoader
 from ai.backend.manager.config.loader.types import AbstractConfigLoader
 from ai.backend.manager.config.provider import ManagerConfigProvider
 from ai.backend.manager.config.watchers.etcd import EtcdConfigWatcher
-from ai.backend.manager.sokovan.deployment.deployment_controller import (
-    DeploymentController,
-    DeploymentControllerArgs,
-)
-from ai.backend.manager.sokovan.deployment.route.route_controller import (
-    RouteController,
-    RouteControllerArgs,
-)
 
 from . import __version__
 from .api.context import RootContext
-from .types import DistributedLockFactory, SMTPTriggerPolicy
+from .providers.agent_registry import agent_registry_ctx
+from .providers.background_task import background_task_ctx
+from .providers.database import database_ctx
+from .providers.distributed_lock import distributed_lock_ctx
+from .providers.event_dispatcher import event_dispatcher_ctx
+from .providers.event_dispatcher_plugin import event_dispatcher_plugin_ctx
+from .providers.event_hub import event_hub_ctx
+from .providers.event_producer import event_producer_ctx
+from .providers.hook_plugin import hook_plugin_ctx
+from .providers.idle_checker import idle_checker_ctx
+from .providers.leader_election import leader_election_ctx
+from .providers.manager_status import manager_status_ctx
+from .providers.message_queue import message_queue_ctx
+from .providers.monitoring import monitoring_ctx
+from .providers.network_plugin import network_plugin_ctx
+from .providers.processors import processors_ctx
+from .providers.redis import redis_ctx
+from .providers.repositories import repositories_ctx
+from .providers.sched_dispatcher import sched_dispatcher_ctx
+from .providers.service_discovery import service_discovery_ctx
+from .providers.services import services_ctx
+from .providers.sokovan_orchestrator import sokovan_orchestrator_ctx
+from .providers.storage_manager import storage_manager_ctx
 
 if TYPE_CHECKING:
-    from ai.backend.manager.reporters.base import AbstractReporter
-
     from .api.types import (
         AppCreator,
         CleanupContext,
@@ -520,725 +481,6 @@ async def webapp_plugin_ctx(root_app: web.Application) -> AsyncIterator[None]:
     await plugin_ctx.cleanup()
 
 
-@actxmgr
-async def manager_status_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .api import ManagerStatus
-
-    if root_ctx.pidx == 0:
-        mgr_status = await root_ctx.config_provider.legacy_etcd_config_loader.get_manager_status()
-        if mgr_status is None or mgr_status not in (ManagerStatus.RUNNING, ManagerStatus.FROZEN):
-            # legacy transition: we now have only RUNNING or FROZEN for HA setup.
-            await root_ctx.config_provider.legacy_etcd_config_loader.update_manager_status(
-                ManagerStatus.RUNNING
-            )
-            mgr_status = ManagerStatus.RUNNING
-        log.info("Manager status: {}", mgr_status)
-        tz = root_ctx.config_provider.config.system.timezone
-        log.info("Configured timezone: {}", tz.tzname(datetime.now()))
-    yield
-
-
-@actxmgr
-async def redis_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    valkey_profile_target = root_ctx.config_provider.config.redis.to_valkey_profile_target()
-    root_ctx.valkey_profile_target = valkey_profile_target
-
-    root_ctx.valkey_container_log = await ValkeyContainerLogClient.create(
-        valkey_profile_target.profile_target(RedisRole.CONTAINER_LOG),
-        db_id=REDIS_CONTAINER_LOG,
-        human_readable_name="container_log",  # saving container_log queue
-    )
-    root_ctx.valkey_live = await ValkeyLiveClient.create(
-        valkey_profile_target.profile_target(RedisRole.LIVE),
-        db_id=REDIS_LIVE_DB,
-        human_readable_name="live",  # tracking live status of various entities
-    )
-    root_ctx.valkey_stat = await ValkeyStatClient.create(
-        valkey_profile_target.profile_target(RedisRole.STATISTICS),
-        db_id=REDIS_STATISTICS_DB,
-        human_readable_name="stat",  # temporary storage for stat snapshots
-    )
-    root_ctx.valkey_image = await ValkeyImageClient.create(
-        valkey_profile_target.profile_target(RedisRole.IMAGE),
-        db_id=REDIS_IMAGE_DB,
-        human_readable_name="image",  # per-agent image availability
-    )
-    root_ctx.valkey_stream = await ValkeyStreamClient.create(
-        valkey_profile_target.profile_target(RedisRole.STREAM),
-        human_readable_name="stream",
-        db_id=REDIS_STREAM_DB,
-    )
-    root_ctx.valkey_schedule = await ValkeyScheduleClient.create(
-        valkey_profile_target.profile_target(RedisRole.STREAM),
-        db_id=REDIS_LIVE_DB,
-        human_readable_name="schedule",  # scheduling marks and coordination
-    )
-    root_ctx.valkey_bgtask = await ValkeyBgtaskClient.create(
-        valkey_profile_target.profile_target(RedisRole.BGTASK),
-        human_readable_name="bgtask",
-        db_id=REDIS_BGTASK_DB,
-    )
-    # Ping ValkeyLiveClient directly
-    await root_ctx.valkey_live.get_server_time()
-    # ValkeyImageClient has its own connection handling
-    # No need to ping it separately as it's already connected
-    yield
-    await root_ctx.valkey_container_log.close()
-    await root_ctx.valkey_image.close()
-    await root_ctx.valkey_stat.close()
-    await root_ctx.valkey_live.close()
-    await root_ctx.valkey_stream.close()
-    await root_ctx.valkey_schedule.close()
-    await root_ctx.valkey_bgtask.close()
-
-
-@actxmgr
-async def database_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .models.utils import connect_database
-
-    async with connect_database(root_ctx.config_provider.config.db) as db:
-        root_ctx.db = db
-        yield
-
-
-def _make_registered_reporters(
-    root_ctx: RootContext,
-) -> dict[str, AbstractReporter]:
-    from .reporters.smtp import SMTPReporter, SMTPSenderArgs
-
-    reporters: dict[str, AbstractReporter] = {}
-    smtp_configs = root_ctx.config_provider.config.reporter.smtp
-    for smtp_conf in smtp_configs:
-        smtp_args = SMTPSenderArgs(
-            host=smtp_conf.host,
-            port=smtp_conf.port,
-            username=smtp_conf.username,
-            password=smtp_conf.password,
-            sender=smtp_conf.sender,
-            recipients=smtp_conf.recipients,
-            use_tls=smtp_conf.use_tls,
-            max_workers=smtp_conf.max_workers,
-            template=smtp_conf.template,
-        )
-        trigger_policy = SMTPTriggerPolicy[smtp_conf.trigger_policy]
-        reporters[smtp_conf.name] = SMTPReporter(smtp_args, trigger_policy)
-
-    return reporters
-
-
-def _make_action_reporters(
-    root_ctx: RootContext,
-    reporters: dict[str, AbstractReporter],
-) -> dict[str, list[AbstractReporter]]:
-    action_monitors: dict[str, list[AbstractReporter]] = {}
-    action_monitor_configs = root_ctx.config_provider.config.reporter.action_monitors
-    for action_monitor_conf in action_monitor_configs:
-        reporter_name: str = action_monitor_conf.reporter
-        try:
-            reporter = reporters[reporter_name]
-        except KeyError:
-            log.warning(f'Invalid Reporter: "{reporter_name}"')
-            continue
-
-        for action_type in action_monitor_conf.subscribed_actions:
-            action_monitors.setdefault(action_type, []).append(reporter)
-
-    return action_monitors
-
-
-@actxmgr
-async def processors_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .actions.monitors.audit_log import AuditLogMonitor
-    from .actions.monitors.prometheus import PrometheusMonitor
-    from .actions.monitors.reporter import ReporterMonitor
-    from .reporters.hub import ReporterHub, ReporterHubArgs
-    from .services.processors import ProcessorArgs, Processors, ServiceArgs
-
-    registered_reporters = _make_registered_reporters(root_ctx)
-    action_reporters = _make_action_reporters(root_ctx, registered_reporters)
-    reporter_hub = ReporterHub(
-        ReporterHubArgs(
-            reporters=action_reporters,
-        )
-    )
-    reporter_monitor = ReporterMonitor(reporter_hub)
-    prometheus_monitor = PrometheusMonitor()
-    audit_log_monitor = AuditLogMonitor(root_ctx.db)
-    root_ctx.processors = Processors.create(
-        ProcessorArgs(
-            service_args=ServiceArgs(
-                db=root_ctx.db,
-                repositories=root_ctx.repositories,
-                etcd=root_ctx.etcd,
-                config_provider=root_ctx.config_provider,
-                storage_manager=root_ctx.storage_manager,
-                valkey_stat_client=root_ctx.valkey_stat,
-                valkey_live=root_ctx.valkey_live,
-                event_fetcher=root_ctx.event_fetcher,
-                background_task_manager=root_ctx.background_task_manager,
-                event_hub=root_ctx.event_hub,
-                agent_registry=root_ctx.registry,
-                error_monitor=root_ctx.error_monitor,
-                idle_checker_host=root_ctx.idle_checker_host,
-                event_dispatcher=root_ctx.event_dispatcher,
-                hook_plugin_ctx=root_ctx.hook_plugin_ctx,
-                scheduling_controller=root_ctx.scheduling_controller,
-                deployment_controller=root_ctx.deployment_controller,
-                event_producer=root_ctx.event_producer,
-                agent_cache=root_ctx.agent_cache,
-            )
-        ),
-        [reporter_monitor, prometheus_monitor, audit_log_monitor],
-    )
-    yield
-
-
-@actxmgr
-async def distributed_lock_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    root_ctx.distributed_lock_factory = init_lock_factory(root_ctx)
-    yield
-
-
-@actxmgr
-async def event_hub_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    root_ctx.event_hub = EventHub()
-    yield
-    await root_ctx.event_hub.shutdown()
-
-
-@actxmgr
-async def service_discovery_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    sd_type = root_ctx.config_provider.config.service_discovery.type
-    match sd_type:
-        case ServiceDiscoveryType.ETCD:
-            root_ctx.service_discovery = ETCDServiceDiscovery(
-                ETCDServiceDiscoveryArgs(root_ctx.etcd)
-            )
-        case ServiceDiscoveryType.REDIS:
-            live_valkey_target = root_ctx.valkey_profile_target.profile_target(RedisRole.LIVE)
-            root_ctx.service_discovery = await RedisServiceDiscovery.create(
-                RedisServiceDiscoveryArgs(valkey_target=live_valkey_target)
-            )
-
-    root_ctx.sd_loop = ServiceDiscoveryLoop(
-        sd_type,
-        root_ctx.service_discovery,
-        ServiceMetadata(
-            display_name=f"manager-{root_ctx.config_provider.config.manager.id}",
-            service_group="manager",
-            version=__version__,
-            endpoint=ServiceEndpoint(
-                address=root_ctx.config_provider.config.manager.announce_addr.address,
-                port=root_ctx.config_provider.config.manager.announce_addr.port,
-                protocol="http",
-                prometheus_address=root_ctx.config_provider.config.manager.announce_internal_addr.address,
-            ),
-        ),
-    )
-
-    if root_ctx.config_provider.config.otel.enabled:
-        meta = root_ctx.sd_loop.metadata
-        otel_spec = OpenTelemetrySpec(
-            service_id=meta.id,
-            service_name=meta.service_group,
-            service_version=meta.version,
-            log_level=root_ctx.config_provider.config.otel.log_level,
-            endpoint=root_ctx.config_provider.config.otel.endpoint,
-        )
-        BraceStyleAdapter.apply_otel(otel_spec)
-    yield
-    root_ctx.sd_loop.close()
-
-
-@actxmgr
-async def message_queue_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    root_ctx.message_queue = await _make_message_queue(root_ctx)
-    yield
-    await root_ctx.message_queue.close()
-
-
-@actxmgr
-async def event_producer_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    root_ctx.event_fetcher = EventFetcher(root_ctx.message_queue)
-    root_ctx.event_producer = EventProducer(
-        root_ctx.message_queue,
-        source=AGENTID_MANAGER,
-        log_events=root_ctx.config_provider.config.debug.log_events,
-    )
-    yield
-    await root_ctx.event_producer.close()
-    await asyncio.sleep(0.2)
-
-
-@actxmgr
-async def event_dispatcher_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .event_dispatcher.dispatch import DispatcherArgs, Dispatchers
-
-    root_ctx.event_dispatcher = EventDispatcher(
-        root_ctx.message_queue,
-        log_events=root_ctx.config_provider.config.debug.log_events,
-        event_observer=root_ctx.metrics.event,
-    )
-    dispatchers = Dispatchers(
-        DispatcherArgs(
-            root_ctx.valkey_container_log,
-            root_ctx.valkey_stat,
-            root_ctx.valkey_stream,
-            root_ctx.scheduler_dispatcher,
-            root_ctx.sokovan_orchestrator.coordinator,
-            root_ctx.scheduling_controller,
-            root_ctx.sokovan_orchestrator.deployment_coordinator,
-            root_ctx.sokovan_orchestrator.route_coordinator,
-            root_ctx.repositories.scheduler.repository,
-            root_ctx.event_hub,
-            root_ctx.registry,
-            root_ctx.db,
-            root_ctx.idle_checker_host,
-            root_ctx.event_dispatcher_plugin_ctx,
-            root_ctx.repositories,
-            lambda: root_ctx.processors,
-            root_ctx.storage_manager,
-            root_ctx.config_provider,
-            use_sokovan=root_ctx.config_provider.config.manager.use_sokovan,
-        )
-    )
-    dispatchers.dispatch(root_ctx.event_dispatcher)
-    await root_ctx.event_dispatcher.start()
-    yield
-    await root_ctx.event_dispatcher.close()
-
-
-async def _make_message_queue(
-    root_ctx: RootContext,
-) -> AbstractMessageQueue:
-    redis_profile_target = root_ctx.config_provider.config.redis.to_redis_profile_target()
-    stream_redis_target = redis_profile_target.profile_target(RedisRole.STREAM)
-    node_id = root_ctx.config_provider.config.manager.id
-    args = RedisMQArgs(
-        anycast_stream_key="events",
-        broadcast_channel="events_all",
-        consume_stream_keys={
-            "events",
-        },
-        subscribe_channels={
-            "events_all",
-        },
-        group_name=EVENT_DISPATCHER_CONSUMER_GROUP,
-        node_id=node_id,
-        db=REDIS_STREAM_DB,
-    )
-    if root_ctx.config_provider.config.manager.use_experimental_redis_event_dispatcher:
-        return HiRedisQueue(
-            stream_redis_target,
-            args,
-        )
-    return await RedisQueue.create(
-        stream_redis_target,
-        args,
-    )
-
-
-@actxmgr
-async def idle_checker_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .idle import init_idle_checkers
-
-    root_ctx.idle_checker_host = await init_idle_checkers(
-        root_ctx.db,
-        root_ctx.config_provider,
-        root_ctx.event_producer,
-        root_ctx.distributed_lock_factory,
-    )
-    await root_ctx.idle_checker_host.start()
-    yield
-    await root_ctx.idle_checker_host.shutdown()
-
-
-@actxmgr
-async def storage_manager_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .models.storage import StorageSessionManager
-
-    root_ctx.storage_manager = StorageSessionManager(root_ctx.config_provider.config.volumes)
-    yield
-    await root_ctx.storage_manager.aclose()
-
-
-@actxmgr
-async def repositories_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .repositories.repositories import Repositories
-    from .repositories.types import RepositoryArgs
-
-    repositories = Repositories.create(
-        args=RepositoryArgs(
-            db=root_ctx.db,
-            storage_manager=root_ctx.storage_manager,
-            config_provider=root_ctx.config_provider,
-            valkey_stat_client=root_ctx.valkey_stat,
-            valkey_live_client=root_ctx.valkey_live,
-            valkey_schedule_client=root_ctx.valkey_schedule,
-            valkey_image_client=root_ctx.valkey_image,
-        )
-    )
-    root_ctx.repositories = repositories
-    yield
-
-
-@actxmgr
-async def network_plugin_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .plugin.network import NetworkPluginContext
-
-    ctx = NetworkPluginContext(
-        root_ctx.etcd,
-        root_ctx.config_provider.config.model_dump(by_alias=True),
-    )
-    root_ctx.network_plugin_ctx = ctx
-    await ctx.init(
-        context=root_ctx,
-        allowlist=root_ctx.config_provider.config.manager.allowed_plugins,
-        blocklist=root_ctx.config_provider.config.manager.disabled_plugins,
-    )
-    log.info("NetworkPluginContext initialized with plugins: {}", list(ctx.plugins.keys()))
-    yield
-    await ctx.cleanup()
-
-
-@actxmgr
-async def hook_plugin_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    ctx = HookPluginContext(
-        root_ctx.etcd,
-        root_ctx.config_provider.config.model_dump(by_alias=True),
-    )
-    root_ctx.hook_plugin_ctx = ctx
-    await ctx.init(
-        context=root_ctx,
-        allowlist=root_ctx.config_provider.config.manager.allowed_plugins,
-        blocklist=root_ctx.config_provider.config.manager.disabled_plugins,
-    )
-    hook_result = await ctx.dispatch(
-        "ACTIVATE_MANAGER",
-        (),
-        return_when=ALL_COMPLETED,
-    )
-    if hook_result.status != PASSED:
-        raise RuntimeError("Could not activate the manager instance.")
-    yield
-    await ctx.cleanup()
-
-
-@actxmgr
-async def event_dispatcher_plugin_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    ctx = EventDispatcherPluginContext(
-        root_ctx.etcd,
-        root_ctx.config_provider.config.model_dump(by_alias=True),
-    )
-    root_ctx.event_dispatcher_plugin_ctx = ctx
-    await ctx.init(
-        context=root_ctx,
-        allowlist=root_ctx.config_provider.config.manager.allowed_plugins,
-        blocklist=root_ctx.config_provider.config.manager.disabled_plugins,
-    )
-    yield
-    await ctx.cleanup()
-
-
-@actxmgr
-async def agent_registry_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from zmq.auth.certs import load_certificate
-
-    from ai.backend.manager.sokovan.scheduling_controller import (
-        SchedulingController,
-        SchedulingControllerArgs,
-    )
-
-    from .agent_cache import AgentRPCCache
-    from .registry import AgentRegistry
-
-    # Create scheduling controller first
-    root_ctx.scheduling_controller = SchedulingController(
-        SchedulingControllerArgs(
-            repository=root_ctx.repositories.scheduler.repository,
-            config_provider=root_ctx.config_provider,
-            storage_manager=root_ctx.storage_manager,
-            event_producer=root_ctx.event_producer,
-            valkey_schedule=root_ctx.valkey_schedule,
-            network_plugin_ctx=root_ctx.network_plugin_ctx,
-        )
-    )
-    # Create deployment controller
-    root_ctx.deployment_controller = DeploymentController(
-        DeploymentControllerArgs(
-            scheduling_controller=root_ctx.scheduling_controller,
-            deployment_repository=root_ctx.repositories.deployment.repository,
-            config_provider=root_ctx.config_provider,
-            storage_manager=root_ctx.storage_manager,
-            event_producer=root_ctx.event_producer,
-            valkey_schedule=root_ctx.valkey_schedule,
-        )
-    )
-    root_ctx.route_controller = RouteController(
-        RouteControllerArgs(
-            valkey_schedule=root_ctx.valkey_schedule,
-        )
-    )
-    manager_pkey, manager_skey = load_certificate(
-        root_ctx.config_provider.config.manager.rpc_auth_manager_keypair
-    )
-    assert manager_skey is not None
-    manager_public_key = PublicKey(manager_pkey)
-    manager_secret_key = SecretKey(manager_skey)
-    root_ctx.agent_cache = AgentRPCCache(root_ctx.db, manager_public_key, manager_secret_key)
-    root_ctx.registry = AgentRegistry(
-        root_ctx.config_provider,
-        root_ctx.db,
-        root_ctx.agent_cache,
-        root_ctx.valkey_stat,
-        root_ctx.valkey_live,
-        root_ctx.valkey_image,
-        root_ctx.event_producer,
-        root_ctx.event_hub,
-        root_ctx.storage_manager,
-        root_ctx.hook_plugin_ctx,
-        root_ctx.network_plugin_ctx,
-        root_ctx.scheduling_controller,
-        debug=root_ctx.config_provider.config.debug.enabled,
-        manager_public_key=manager_public_key,
-        manager_secret_key=manager_secret_key,
-        use_sokovan=root_ctx.config_provider.config.manager.use_sokovan,
-    )
-    await root_ctx.registry.init()
-    yield
-    await root_ctx.registry.shutdown()
-
-
-@actxmgr
-async def sched_dispatcher_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .scheduler.dispatcher import SchedulerDispatcher
-
-    root_ctx.scheduler_dispatcher = await SchedulerDispatcher.create(
-        root_ctx.config_provider,
-        root_ctx.etcd,
-        root_ctx.event_producer,
-        root_ctx.distributed_lock_factory,
-        root_ctx.registry,
-        root_ctx.valkey_live,
-        root_ctx.valkey_stat,
-        root_ctx.repositories.schedule.repository,
-    )
-    yield
-    await root_ctx.scheduler_dispatcher.close()
-
-
-@actxmgr
-async def leader_election_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    """Initialize leader election for distributed coordination."""
-    import socket
-
-    from ai.backend.common.clients.valkey_client.valkey_leader.client import ValkeyLeaderClient
-    from ai.backend.common.leader import ValkeyLeaderElection, ValkeyLeaderElectionConfig
-    from ai.backend.common.leader.tasks import EventProducerTask, LeaderCron, PeriodicTask
-
-    # Create ValkeyLeaderClient for leader election
-    valkey_leader_client = await ValkeyLeaderClient.create(
-        valkey_target=root_ctx.valkey_profile_target.profile_target(RedisRole.STREAM),
-        db_id=REDIS_STREAM_LOCK,  # Use a dedicated DB for leader election
-        human_readable_name="leader",
-    )
-
-    # Create leader election configuration
-    server_id = f"manager-{socket.gethostname()}-{root_ctx.pidx}"
-    leader_config = ValkeyLeaderElectionConfig(
-        server_id=server_id,
-        leader_key="leader:sokovan:scheduler",
-        lease_duration=30,
-        renewal_interval=10.0,
-        failure_threshold=3,
-    )
-
-    # Create leader election instance
-    root_ctx.leader_election = ValkeyLeaderElection(
-        leader_client=valkey_leader_client,
-        config=leader_config,
-    )
-
-    # Get task specifications from sokovan and register them
-    task_specs = root_ctx.sokovan_orchestrator.create_task_specs()
-
-    # Rescan reservoir registry periodically
-    task_specs.append(
-        EventTaskSpec(
-            name="reservoir_registry_scan",
-            event_factory=lambda: DoScanReservoirRegistryEvent(),
-            interval=3600,  # 1 hour
-            initial_delay=0,
-        )
-    )
-
-    # Create event producer tasks from specs
-    leader_tasks: list[PeriodicTask] = [
-        EventProducerTask(spec, root_ctx.event_producer) for spec in task_specs
-    ]
-
-    # Register tasks with the election system
-    leader_cron = LeaderCron(tasks=leader_tasks)
-    root_ctx.leader_election.register_task(leader_cron)
-
-    # Start leader election (will start tasks when becoming leader)
-    await root_ctx.leader_election.start()
-    log.info(f"Leader election started for server {server_id}")
-
-    yield
-
-    # Cleanup leader election
-    await root_ctx.leader_election.stop()
-
-
-@actxmgr
-async def sokovan_orchestrator_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .clients.agent import AgentPool
-    from .sokovan.scheduler.factory import create_default_scheduler
-    from .sokovan.sokovan import SokovanOrchestrator
-
-    # Create agent pool for scheduler
-    agent_pool = AgentPool(root_ctx.agent_cache)
-
-    # Create scheduler with default components
-    scheduler = create_default_scheduler(
-        root_ctx.repositories.scheduler.repository,
-        root_ctx.repositories.deployment.repository,
-        root_ctx.config_provider,
-        root_ctx.distributed_lock_factory,
-        agent_pool,
-        root_ctx.network_plugin_ctx,
-        root_ctx.event_producer,
-        root_ctx.valkey_schedule,
-    )
-
-    # Create HTTP client pool for deployment operations
-    from ai.backend.common.clients.http_client.client_pool import (
-        ClientPool,
-        tcp_client_session_factory,
-    )
-    from ai.backend.manager.sokovan.deployment.coordinator import DeploymentCoordinator
-    from ai.backend.manager.sokovan.deployment.route.coordinator import RouteCoordinator
-
-    client_pool = ClientPool(tcp_client_session_factory)
-
-    # Create deployment coordinator
-    deployment_coordinator = DeploymentCoordinator(
-        valkey_schedule=root_ctx.valkey_schedule,
-        deployment_controller=root_ctx.deployment_controller,
-        deployment_repository=root_ctx.repositories.deployment.repository,
-        event_producer=root_ctx.event_producer,
-        lock_factory=root_ctx.distributed_lock_factory,
-        config_provider=root_ctx.config_provider,
-        scheduling_controller=root_ctx.scheduling_controller,
-        client_pool=client_pool,
-        valkey_stat=root_ctx.valkey_stat,
-        route_controller=root_ctx.route_controller,
-    )
-
-    # Create route coordinator
-    route_coordinator = RouteCoordinator(
-        valkey_schedule=root_ctx.valkey_schedule,
-        deployment_repository=root_ctx.repositories.deployment.repository,
-        event_producer=root_ctx.event_producer,
-        lock_factory=root_ctx.distributed_lock_factory,
-        config_provider=root_ctx.config_provider,
-        scheduling_controller=root_ctx.scheduling_controller,
-        client_pool=client_pool,
-    )
-
-    # Create sokovan orchestrator with lock factory for timers
-    root_ctx.sokovan_orchestrator = SokovanOrchestrator(
-        scheduler=scheduler,
-        event_producer=root_ctx.event_producer,
-        valkey_schedule=root_ctx.valkey_schedule,
-        lock_factory=root_ctx.distributed_lock_factory,
-        scheduling_controller=root_ctx.scheduling_controller,
-        deployment_coordinator=deployment_coordinator,
-        route_coordinator=route_coordinator,
-    )
-
-    log.info("Sokovan orchestrator initialized")
-
-    try:
-        yield
-    finally:
-        # Leader election will handle task cleanup
-        pass
-
-
-@actxmgr
-async def monitoring_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .plugin.monitor import ManagerErrorPluginContext, ManagerStatsPluginContext
-
-    ectx = ManagerErrorPluginContext(
-        root_ctx.etcd, root_ctx.config_provider.config.model_dump(by_alias=True)
-    )
-    sctx = ManagerStatsPluginContext(
-        root_ctx.etcd, root_ctx.config_provider.config.model_dump(by_alias=True)
-    )
-    init_success = False
-
-    try:
-        await ectx.init(
-            context={"_root.context": root_ctx},
-            allowlist=root_ctx.config_provider.config.manager.allowed_plugins,
-        )
-        await sctx.init(allowlist=root_ctx.config_provider.config.manager.allowed_plugins)
-    except Exception:
-        log.error("Failed to initialize monitoring plugins")
-    else:
-        init_success = True
-        root_ctx.error_monitor = ectx
-        root_ctx.stats_monitor = sctx
-    yield
-    if init_success:
-        await sctx.cleanup()
-        await ectx.cleanup()
-
-
-@actxmgr
-async def services_ctx(root_ctx: RootContext) -> AsyncIterator[None]:
-    from .service.base import ServicesContext
-    from .service.container_registry.base import PerProjectRegistryQuotaRepository
-    from .service.container_registry.harbor import (
-        PerProjectContainerRegistryQuotaClientPool,
-        PerProjectContainerRegistryQuotaService,
-    )
-
-    db = root_ctx.db
-
-    per_project_container_registries_quota = PerProjectContainerRegistryQuotaService(
-        repository=PerProjectRegistryQuotaRepository(db),
-        client_pool=PerProjectContainerRegistryQuotaClientPool(),
-    )
-
-    root_ctx.services_ctx = ServicesContext(
-        per_project_container_registries_quota,
-    )
-    yield None
-
-
-class background_task_ctx:
-    def __init__(self, root_ctx: RootContext) -> None:
-        self.root_ctx = root_ctx
-
-    async def __aenter__(self) -> None:
-        self.root_ctx.background_task_manager = BackgroundTaskManager(
-            self.root_ctx.event_producer,
-            valkey_client=self.root_ctx.valkey_bgtask,
-            server_id=self.root_ctx.config_provider.config.manager.id,
-            bgtask_observer=self.root_ctx.metrics.bgtask,
-        )
-
-    async def __aexit__(self, *exc_info) -> None:
-        pass
-
-    async def shutdown(self) -> None:
-        if hasattr(self.root_ctx, "background_task_manager"):
-            await self.root_ctx.background_task_manager.shutdown()
-
-
 def handle_loop_error(
     root_ctx: RootContext,
     loop: asyncio.AbstractEventLoop,
@@ -1285,51 +527,6 @@ def init_subapp(pkg_name: str, root_app: web.Application, create_subapp: AppCrea
     root_ctx: RootContext = root_app["_root.context"]
     subapp, global_middlewares = create_subapp(root_ctx.cors_options)
     _init_subapp(pkg_name, root_app, subapp, global_middlewares)
-
-
-def init_lock_factory(root_ctx: RootContext) -> DistributedLockFactory:
-    ipc_base_path = root_ctx.config_provider.config.manager.ipc_base_path
-    manager_id = root_ctx.config_provider.config.manager.id
-    lock_backend = root_ctx.config_provider.config.manager.distributed_lock
-    log.debug("using {} as the distributed lock backend", lock_backend)
-    match lock_backend:
-        case "filelock":
-            from ai.backend.common.lock import FileLock
-
-            return lambda lock_id, lifetime_hint: FileLock(
-                ipc_base_path / f"{manager_id}.{lock_id}.lock",
-                timeout=0,
-            )
-        case "pg_advisory":
-            from .pglock import PgAdvisoryLock
-
-            return lambda lock_id, lifetime_hint: PgAdvisoryLock(root_ctx.db, lock_id)
-        case "redlock":
-            from ai.backend.common.lock import RedisLock
-
-            redlock_config = root_ctx.config_provider.config.manager.redlock_config
-            redis_profile_target = root_ctx.config_provider.config.redis.to_redis_profile_target()
-            redis_lock = redis_helper.get_redis_object(
-                redis_profile_target.profile_target(RedisRole.STREAM_LOCK),
-                name="lock",  # distributed locks
-                db=REDIS_STREAM_LOCK,
-            )
-            return lambda lock_id, lifetime_hint: RedisLock(
-                str(lock_id),
-                redis_lock,
-                lifetime=min(lifetime_hint * 2, lifetime_hint + 30),
-                lock_retry_interval=redlock_config["lock_retry_interval"],
-            )
-        case "etcd":
-            from ai.backend.common.lock import EtcdLock
-
-            return lambda lock_id, lifetime_hint: EtcdLock(
-                str(lock_id),
-                root_ctx.etcd,
-                lifetime=min(lifetime_hint * 2, lifetime_hint + 30),
-            )
-        case other:
-            raise ValueError(f"Invalid lock backend: {other}")
 
 
 def build_root_app(
